Projects/WebsiteChatbot/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
anthropic>=0.84.0
sentence-transformers>=2.7.0
beautifulsoup4>=4.11.1
lxml>=4.9.0
python-dotenv>=1.0.0
numpy>=1.24.0
slowapi>=0.1.9
